+++ options/db_options.h
@@ -122,7 +122,6 @@ struct MutableDBOptions {
   void Dump(Logger* log) const;
 
   int max_background_jobs;
-  int base_background_compactions;
   int max_background_compactions;
   uint32_t max_subcompactions;
   bool avoid_flush_during_shutdown;